requirements.txt
attrs==17.3.0
Django==2.0
djangorestframework==3.7.3
pluggy==0.6.0
py==1.5.2
pytest==3.3.1
pytest-django==3.1.2
pytz==2017.3
six==1.11.0
django-cors-headers==2.1.0
